Rename : Change Components Directory - 컴포넌트 폴더구조가 보기 지저분한 부분이 있어서 전체를 변경했습니다. - 이제 페이지에 해당하는 건 최대한 페이지에만 두고 그 이외의 것들은   컴포넌트 폴더에 두는 걸로 하면 될 듯 합니다..!
