+++ frontend/src/components/avatar/AvatarIcon.tsx
