+++ frontend/src/components/shadcn/avatar.tsx
