+++ frontend/src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import Layout from '@/components/components/custom/Layout';
+import Layout from '@/components/Layout';
 import type {AppProps} from 'next/app';
 import '../../app/globals.css';
 export default function App({Component}: AppProps) {
